requirements.txt
boto3
transformers
torch
matplotlib
pillow
scikit-learn
opencv-contrib-python
pytesseract
easyocr
pandas